Added option to make docker image unhealthy by adding file /app/unhealthy

+++ healthcheck.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
+if [ -f /app/unhealthy ]; then
+  exit 1
+fi
+
 STATUS=$(curl -s -o /dev/null -w "%{http_code}" -f -I -m 4 -A "HEALTHCHECK" http://localhost/)
 
 if [ $STATUS -eq 200 ]; then
